requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.1.0
Markdown==3.4.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
zipp==3.15.0
